fix - ajuste para impressao de long int com %u

+++ ft_printf.c
@@ -6,7 +6,7 @@
 /*   By: evportel <evportel@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/14 11:47:08 by evportel          #+#    #+#             */
-/*   Updated: 2023/06/20 12:06:25 by evportel         ###   ########.fr       */
+/*   Updated: 2023/06/20 17:41:02 by evportel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -46,7 +46,7 @@ static void	show_of_parameter(char charactere, va_list arg)
 	else if (charactere == 'd' || charactere == 'i')
 		ft_putnbr(va_arg(arg, int));
 	else if (charactere == 'u')
-		ft_putnbr_unsigned(va_arg(arg, long int));
+		ft_putnbr_unsigned(va_arg(arg, unsigned int));
 }
 
 int	ft_printf(const char *format, ...)
